// src/shared/api/generated/model/productSubscriptionExecuteResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { ProductSubscriptionExecuteResponseProductGroup } from './productSubscriptionExecuteResponseProductGroup';
import type { ProductSubscriptionExecuteResponseStatus } from './productSubscriptionExecuteResponseStatus';

export interface ProductSubscriptionExecuteResponse {
  subscriptionId?: number;
  accountId?: number;
  accountNumber?: string;
  productName?: string;
  productGroup?: ProductSubscriptionExecuteResponseProductGroup;
  subscriptionAmount?: number;
  termMonths?: number;
  appliedRate?: number;
  openedDate?: string;
  maturityDate?: string;
  expectedMaturityAmount?: number;
  status?: ProductSubscriptionExecuteResponseStatus;
  transactionNumber?: string;
  subscribedAt?: string;
}