// src/shared/api/generated/model/productSubscriptionResultResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { AutoTransferPrefill } from './autoTransferPrefill';
import type { ProductSubscriptionResultResponseProductGroup } from './productSubscriptionResultResponseProductGroup';

export interface ProductSubscriptionResultResponse {
  subscriptionId?: number;
  accountId?: number;
  accountNumber?: string;
  productId?: number;
  productName?: string;
  productGroup?: ProductSubscriptionResultResponseProductGroup;
  subscriptionAmount?: number;
  termMonths?: number;
  appliedRate?: number;
  openedDate?: string;
  maturityDate?: string;
  expectedMaturityAmount?: number;
  status?: string;
  transactionNumber?: string;
  subscribedAt?: string;
  autoTransferPrefill?: AutoTransferPrefill;
}